crowd_server/requirements/base.txt
dj-database-url==1.0.0
Django==4.1
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.2.1
djangorestframework==3.13.1
Markdown==3.4.1
django-filter==22.1
django-rest-swagger==2.2.0
django-cors-headers==3.13.0